Refactor: 	rename of a directory: src/userdb/user -> src/userdb/storage
