+++ oauth2_passkey/src/userdb/mod.rs
@@ -1,11 +1,11 @@
 mod errors;
+mod storage;
 mod types;
-mod user;
 
 pub use errors::UserError;
+pub(crate) use storage::DB_TABLE_USERS;
+pub use storage::UserStore;
 pub use types::User;
-pub(crate) use user::DB_TABLE_USERS;
-pub use user::UserStore;
 
 pub async fn init() -> Result<(), UserError> {
     UserStore::init().await
